refactor(experimental): update paths to commitment type
