+++ packages/rpc-core/src/rpc-methods/__tests__/get-vote-accounts.ts
@@ -3,7 +3,8 @@ import { createHttpTransport, createJsonRpc } from '@solana/rpc-transport';
 import type { Rpc } from '@solana/rpc-transport/dist/types/json-rpc-types';
 import fetchMock from 'jest-fetch-mock-fork';
 
-import { Commitment, createSolanaRpcApi, SolanaRpcMethods } from '../index';
+import { Commitment } from '../../commitment';
+import { createSolanaRpcApi, SolanaRpcMethods } from '../index';
 
 describe('getVoteAccounts', () => {
     let rpc: Rpc<SolanaRpcMethods>;
